--- src/main/java/kr/hhplus/be/server/interfaces/order/dto/request/OrderRequest.java
@@ -1,21 +1,0 @@
-package kr.hhplus.be.server.interfaces.order.dto.request;
-
-import kr.hhplus.be.server.domain.product.Product;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.antlr.v4.runtime.misc.NotNull;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-public class OrderRequest {
-
-    @NotNull
-    private Long userId;
-
-    @NotNull
-    private List<Product> orderItmes;
-
-    private Long couponId;
-}
