--- src/main/java/kr/hhplus/be/server/interfaces/point/dto/request/PointRequest.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.interfaces.point.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.antlr.v4.runtime.misc.NotNull;
-
-@Getter
-@NoArgsConstructor
-public class PointRequest {
-
-    @NotNull
-    private Long userId;
-
-    @NotNull
-    private Long chargeAmount;
-}
