+++ src/main/java/kr/hhplus/be/server/interfaces/web/coupon/dto/request/CouponRequest.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.interfaces.web.coupon.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.antlr.v4.runtime.misc.NotNull;
+
+@Getter
+@NoArgsConstructor
+public class CouponRequest {
+
+    @NotNull
+    private Long userId;
+
+    @NotNull
+    private Long couponId;
+}
