+++ src/main/java/kr/hhplus/be/server/interfaces/web/coupon/dto/response/CouponIssueResponse.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.interfaces.web.coupon.dto.response;
+
+import kr.hhplus.be.server.domain.coupon.CouponIssuance;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CouponIssueResponse {
+    private String message;
+    private CouponIssuance data;
+}
