+++ src/main/java/kr/hhplus/be/server/interfaces/web/coupon/dto/response/UserCouponResponse.java
@@ -1,0 +1,18 @@
+package kr.hhplus.be.server.interfaces.web.coupon.dto.response;
+
+import kr.hhplus.be.server.domain.coupon.CouponIssuance;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserCouponResponse {
+    private String message;
+    private List<CouponIssuance> data;
+}
