+++ src/main/java/kr/hhplus/be/server/interfaces/web/order/dto/response/OrderResponse.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.interfaces.web.order.dto.response;
+
+import kr.hhplus.be.server.domain.order.Order;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderResponse {
+    private String message;
+    private Order data;
+}
