+++ src/main/java/kr/hhplus/be/server/interfaces/web/payment/controller/PaymentController.java
@@ -1,0 +1,22 @@
+package kr.hhplus.be.server.interfaces.web.payment.controller;
+
+import kr.hhplus.be.server.domain.order.OrderPay;
+import kr.hhplus.be.server.interfaces.web.ApiResponse;
+import kr.hhplus.be.server.interfaces.web.payment.dto.request.PaymentRequest;
+import kr.hhplus.be.server.interfaces.web.payment.dto.response.PaymentResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/payments")
+public class PaymentController {
+
+    @PostMapping
+    public ApiResponse<PaymentResponse> payOrder(@RequestBody PaymentRequest request) {
+        return ApiResponse.ok(new PaymentResponse("결제 성공", new OrderPay(1L, 1L, 15000L)));
+    }
+}
