+++ src/main/java/kr/hhplus/be/server/interfaces/web/payment/dto/response/PaymentResponse.java
@@ -1,0 +1,17 @@
+package kr.hhplus.be.server.interfaces.web.payment.dto.response;
+
+import kr.hhplus.be.server.domain.order.OrderPay;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PaymentResponse {
+
+    private String message;
+    private OrderPay data;
+}
