+++ src/main/java/kr/hhplus/be/server/interfaces/web/product/dto/response/ProductResponse.java
@@ -1,0 +1,18 @@
+package kr.hhplus.be.server.interfaces.web.product.dto.response;
+
+import kr.hhplus.be.server.domain.product.Product;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductResponse {
+    private String message;
+    private List<Product> data;
+}
